Refactor CommandResult to be more extension-friendly.
