+++ src/main/java/seedu/mark/logic/commands/Command.java
@@ -1,5 +1,6 @@
 package seedu.mark.logic.commands;
 
+import seedu.mark.logic.commands.commandresult.CommandResult;
 import seedu.mark.logic.commands.exceptions.CommandException;
 import seedu.mark.model.Model;
