+++ src/main/java/seedu/mark/logic/commands/DeleteCommand.java
@@ -6,6 +6,7 @@ import java.util.List;
 
 import seedu.mark.commons.core.Messages;
 import seedu.mark.commons.core.index.Index;
+import seedu.mark.logic.commands.commandresult.CommandResult;
 import seedu.mark.logic.commands.exceptions.CommandException;
 import seedu.mark.model.Model;
 import seedu.mark.model.bookmark.Bookmark;
