+++ src/main/java/seedu/mark/logic/commands/EditCommand.java
@@ -16,6 +16,7 @@ import java.util.Set;
 import seedu.mark.commons.core.Messages;
 import seedu.mark.commons.core.index.Index;
 import seedu.mark.commons.util.CollectionUtil;
+import seedu.mark.logic.commands.commandresult.CommandResult;
 import seedu.mark.logic.commands.exceptions.CommandException;
 import seedu.mark.model.Model;
 import seedu.mark.model.bookmark.Bookmark;
