+++ src/main/java/seedu/mark/logic/commands/HelpCommand.java
@@ -1,5 +1,6 @@
 package seedu.mark.logic.commands;
 
+import seedu.mark.logic.commands.commandresult.CommandResult;
 import seedu.mark.model.Model;
 
 /**
@@ -16,6 +17,6 @@ public class HelpCommand extends Command {
 
     @Override
     public CommandResult execute(Model model) {
-        return new CommandResult(SHOWING_HELP_MESSAGE, true, false);
+        return new CommandResult(SHOWING_HELP_MESSAGE);
     }
 }
