+++ src/test/java/seedu/mark/logic/LogicManagerTest.java
@@ -17,8 +17,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import seedu.mark.logic.commands.AddCommand;
-import seedu.mark.logic.commands.CommandResult;
 import seedu.mark.logic.commands.ListCommand;
+import seedu.mark.logic.commands.commandresult.CommandResult;
 import seedu.mark.logic.commands.exceptions.CommandException;
 import seedu.mark.logic.parser.exceptions.ParseException;
 import seedu.mark.model.Model;
