+++ src/test/java/seedu/mark/logic/commands/CommandTestUtil.java
@@ -13,6 +13,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import seedu.mark.commons.core.index.Index;
+import seedu.mark.logic.commands.commandresult.CommandResult;
 import seedu.mark.logic.commands.exceptions.CommandException;
 import seedu.mark.model.Mark;
 import seedu.mark.model.Model;
