+++ src/test/java/seedu/mark/logic/commands/ExitCommandTest.java
@@ -5,6 +5,8 @@ import static seedu.mark.logic.commands.ExitCommand.MESSAGE_EXIT_ACKNOWLEDGEMENT
 
 import org.junit.jupiter.api.Test;
 
+import seedu.mark.logic.commands.commandresult.CommandResult;
+import seedu.mark.logic.commands.commandresult.ExitCommandResult;
 import seedu.mark.model.Model;
 import seedu.mark.model.ModelManager;
 
@@ -14,7 +16,7 @@ public class ExitCommandTest {
 
     @Test
     public void execute_exit_success() {
-        CommandResult expectedCommandResult = new CommandResult(MESSAGE_EXIT_ACKNOWLEDGEMENT, false, true);
+        CommandResult expectedCommandResult = new ExitCommandResult(MESSAGE_EXIT_ACKNOWLEDGEMENT);
         assertCommandSuccess(new ExitCommand(), model, expectedCommandResult, expectedModel);
     }
 }
