+++ src/test/java/seedu/mark/logic/commands/HelpCommandTest.java
@@ -5,6 +5,8 @@ import static seedu.mark.logic.commands.HelpCommand.SHOWING_HELP_MESSAGE;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.mark.logic.commands.commandresult.CommandResult;
+import seedu.mark.logic.commands.commandresult.HelpCommandResult;
 import seedu.mark.model.Model;
 import seedu.mark.model.ModelManager;
 
@@ -14,7 +16,7 @@ public class HelpCommandTest {
 
     @Test
     public void execute_help_success() {
-        CommandResult expectedCommandResult = new CommandResult(SHOWING_HELP_MESSAGE, true, false);
+        CommandResult expectedCommandResult = new HelpCommandResult(SHOWING_HELP_MESSAGE);
         assertCommandSuccess(new HelpCommand(), model, expectedCommandResult, expectedModel);
     }
 }
